+++ NEWS.md
@@ -1,3 +1,14 @@
+# rjdqa 0.1.3
+
+- deprecated function `sa_dashboard()` removed.
+
+- legend in `simple_dashboard()`.
+
+- new parameter in `simple_dashboard()` to remove the "Others" contribution in the variance decomposition table and to rescale the contributions to sum to 100.
+
+- new function `simple_dashboard2()`: same as `simple_dashboard()` but with smaller text to add informations on last outliers.
+
+
 # rjdqa 0.1.2
 
 - documentation updated.